feat(workspace): restructure server into discrete services
