+++ libs/routers_server/buf.yaml
@@ -2,3 +2,6 @@ version: v2
 modules:
   - path: proto
     name: buf.build/routers/api
+    includes:
+      - proto/api
+      - proto/model
