+++ libs/routers_server/proto/api/optimise/v1/definition.proto
@@ -1,0 +1,18 @@
+syntax = "proto3";
+package api.optimise.v1;
+
+import "model/v1/costing.proto";
+import "model/v1/geo.proto";
+
+// The request message containing the user's name.
+message RouteRequest {
+  model.v1.Coordinate start = 1;
+  model.v1.Coordinate end = 2;
+  model.v1.Costing costing_method = 3;
+}
+
+// The response message including pathing, and weighted heuristics
+message RouteResponse {
+  repeated model.v1.Coordinate shape = 1;
+  uint32 cost = 2;
+}
