+++ libs/routers_server/proto/api/proximity/v1/definition.proto
@@ -1,0 +1,22 @@
+syntax = "proto3";
+package api.proximity.v1;
+
+import "model/v1/geo.proto";
+
+message ClosestPointRequest {
+  model.v1.Coordinate coordinate = 1;
+}
+
+message ClosestPointResponse {
+  model.v1.Coordinate coordinate = 1;
+}
+
+// Specifies the distance and a position from which to search
+message ClosestSnappedPointRequest {
+  model.v1.Coordinate point = 1;
+  double search_radius = 2;
+}
+
+message ClosestSnappedPointResponse {
+  model.v1.Coordinate coordinate = 1;
+}
