+++ libs/routers_server/proto/api/proximity/v1/service.proto
@@ -1,16 +1,10 @@
 syntax = "proto3";
-package api.v1.router.v1;
+package api.proximity.v1;
 
-import "api/v1/router/v1/definition.proto";
+import "api/proximity/v1/definition.proto";
 
 // The routing service
-service RouterService {
-  // Returns the appropriate route between the starting and ending locations.
-  rpc Route(RouteRequest) returns (RouteResponse);
-
-  // Matches an input route to a given underlying map.
-  rpc MapMatch(MapMatchRequest) returns (MapMatchResponse);
-
+service ProximityService {
   // Provides the coordinate of the closest point on the map to the input
   rpc ClosestPoint(ClosestPointRequest) returns (ClosestPointResponse);
