+++ libs/routers_server/proto/model/v1/costing.proto
@@ -1,0 +1,7 @@
+syntax = "proto3";
+package model.v1;
+
+enum Costing {
+  COSTING_UNSPECIFIED = 0;
+  COSTING_CAR = 1;
+}
