+++ libs/routers_server/src/services/mod.rs
@@ -1,0 +1,27 @@
+use routers::Graph;
+use routers::transition::PredicateCache;
+use std::path::Path;
+use std::sync::{Arc, Mutex};
+
+pub mod matcher;
+pub mod optimise;
+pub mod proximity;
+
+#[derive(Debug)]
+pub struct RouteService {
+    pub graph: Graph,
+    lookup: Arc<Mutex<PredicateCache>>,
+}
+
+impl RouteService {
+    pub fn from_file(file: &str) -> Result<RouteService, Box<dyn std::error::Error>> {
+        let path = Path::new(file);
+        let graph =
+            Graph::new(path.as_os_str().to_ascii_lowercase()).map_err(|e| format!("{:?}", e))?;
+
+        Ok(RouteService {
+            graph,
+            lookup: Arc::new(Mutex::new(PredicateCache::default())),
+        })
+    }
+}
